Add files via upload
Object Types (Blueprints) (Classes)Sometimes we need a "blueprint" for creating many objects of the same "type". The way to create an "object type", is to use an object constructor function .Objects of the same type are created by calling the constructor function with the new keyword. In JavaScript, the thing called this is the object that "owns" the code. The value of this , when used in an object, is the object itself. In a constructor function this does not have a value. It is a substitute for the new object.
Note: It is considered a good practice to capitalize the first letter of your constructor function.
The constructor property returns a reference to the Object constructor function that created the instance object. Note that the value of this property is a reference to the function itself, not a string containing the function's name.